backend/alf/requirements.txt
fastapi
uvicorn[standard]
pydantic
anthropic
numpy
requests
openai
python-multipart
python-dotenv
sqlalchemy
pytest
httpx
